+++ back-end/service/matches.service.ts
@@ -1,6 +1,13 @@
 import matchesDb from "../repository/matches.db"
 
 const getPlayersByTeamAndMatch = async ( matchId: number, teamName: string) => {
+    if (!matchId) {
+        throw new Error('Match Id is required');
+    }
+    if (!teamName) {
+        throw new Error('Team Name is required');
+    }
+
     try {
         const players = await matchesDb.getPlayersByTeamAndMatch(matchId, teamName);
         return players;
@@ -11,6 +18,12 @@ const getPlayersByTeamAndMatch = async ( matchId: number, teamName: string) => {
 }
 
 const addMatches = async (matches: any, eventId: number) => {
+    if (!matches) {
+        throw new Error('Matches are required');
+    }
+    if (!eventId) {
+        throw new Error('Event Id is required');
+    }
     try {
     const newMatches = await matchesDb.addMatches(matches, eventId);
     return newMatches;
@@ -21,6 +34,15 @@ const addMatches = async (matches: any, eventId: number) => {
 }
 
 const editMatches = async (matches: any, eventId: number, matchesId: number) => {
+    if (!matches) {
+        throw new Error('Matches are required');
+    }
+    if (!eventId) {
+        throw new Error('Event Id is required');
+    }
+    if (!matchesId) {
+        throw new Error('Matches Id is required');
+    }
     try {
     const newMatches = await matchesDb.editMatches(matches, eventId, matchesId);
     return newMatches;
@@ -31,6 +53,9 @@ const editMatches = async (matches: any, eventId: number, matchesId: number) => 
 }
 
 const deleteMatches = async (matchesId: number) => {
+    if (!matchesId) {
+        throw new Error('Matches Id is required');
+    }
     try {
         const deletedMatches = await matchesDb.deleteMatches(matchesId);
         return deletedMatches;
@@ -41,6 +66,9 @@ const deleteMatches = async (matchesId: number) => {
 }
 
 const getEventNameByMatch = async (matchId: number) => {
+    if (!matchId) {
+        throw new Error('Match Id is required');
+    }
     const event = await matchesDb.getEventNameByMatch(matchId);
     return event;
 }
